feat: create LeaderBoard feature
This commit creates a new screen called Leaderboard that lists the user of the household that stole the most food and let the most food go expired. Also list the next 4 top "rats" and "stinkers"
This includes making the "king" have a dedicated section of the screen with a specific image related to the leaderboard behind them. Add a button that only appears for the leader that once clicked begins to play themed music (click again to amke it stop).

+++ app/src/main/java/com/android/shelfLife/model/newhousehold/HouseHold.kt
@@ -13,6 +13,6 @@ data class HouseHold(
     var name: String, // Name of the household
     val members: List<String>, // List of member IDs in the household
     val sharedRecipes: List<String>, // List of shared recipe IDs
-    val ratPoints: Map<String, Int>, // Map of member IDs to their respective rat points
-    val stinkyPoints : Map<String, Int> // Map of member IDs to their respective stinky points
+    val ratPoints: Map<String, Long>, // Map of member IDs to their respective rat points
+    val stinkyPoints : Map<String, Long> // Map of member IDs to their respective stinky points
 )
